TST: reorganized LBFGS tests
